V4.0 (#88)
* language arg for syntax highlighting

* unmute fix

* purge fix

* quote fix

* purge fix

+++ modules/purge.py
@@ -34,8 +34,10 @@ async def purge(client: Client, message: Message):
     chunk = []
     async for msg in client.get_chat_history(
         chat_id=message.chat.id,
-        offset_id=message.reply_to_message.id,
+        limit=message.id - message.reply_to_message.id + 1,
     ):
+        if msg.id < message.reply_to_message.id:
+            break
         chunk.append(msg.id)
         if len(chunk) >= 100:
             await client.delete_messages(message.chat.id, chunk)
